requirements.txt
torch
torchvision
transformers
numpy
tqdm
Pillow
requests
huggingface-hub
Flask